+++ CHANGELOG.rst
@@ -8,6 +8,20 @@ Version 0.10 (unreleased)
 
 * Addition of ``nearest`` and ``nearest_all`` functions to ``STRtree`` for
   GEOS >= 3.6 to find the nearest neighbors (#272).
+* Enable bulk construction of geometries with different number of coordinates
+  by optionally taking index arrays in the constructors ``multipoints``,
+  ``points``, ``linestrings``, and ``linearrings`` (#322).
+* Enable bulk construction of collections with different number of geometries
+  by optionally taking index arrays in the constructors ``multipoints``,
+  ``multilinestrings``, ``multipolygons``, and ``geometrycollections`` (#290).
+* Released GIL for ``points``, ``linestrings``, ``linearrings``, and
+  ``polygons`` (without holes) (#310).
+* Added the option to return the geometry index in ``get_coordinates`` (#318).
+* Updated ``box`` ufunc to use internal C function for creating polygon
+  (about 2x faster) and added ``ccw`` parameter to create polygon in
+  counterclockwise (default) or clockwise direction (#308).
+* Added ``to_shapely`` and improved performance of ``from_shapely`` in the case
+  GEOS versions are different (#312).
 
 **API Changes**
 
@@ -16,16 +30,23 @@ Version 0.10 (unreleased)
 * Deprecated ``VALID_PREDICATES`` set from ``pygeos.strtree`` package; these can be constructed
   in downstream libraries using the ``pygeos.strtree.BinaryPredicate`` enum.
   This will be removed in a future release.
+* ``points``, ``linestrings``, ``linearrings``, and ``polygons`` now return a ``GEOSException``
+  instead of a ``ValueError`` for invalid input (#310).
+* Addition of ``on_invalid`` parameter to ``from_wkb`` and ``from_wkt`` to
+  optionally return invalid WKB geometries as ``None``.
 
 **Added GEOS functions**
 
 * Addition of a ``contains_properly`` function (#267)
 * Addition of a ``polygonize`` function (#275)
+* Addition of a ``polygonize_full`` function (#298)
 * Addition of a ``segmentize`` function for GEOS >= 3.10 (#299)
 
 **Bug fixes**
 
 * Fixed portability issue for ARM architecture (#293)
+* Fixed segfault in ``linearrings`` and ``box`` when constructing a geometry with nan
+  coordinates (#310).
 
 **Acknowledgments**
 
@@ -35,6 +56,7 @@ People with a "+" by their names contributed a patch for the first time.
 * Brendan Ward
 * Casper van der Wel
 * Joris Van den Bossche
+* 0phoff +
 
 
 Version 0.9 (2021-01-23)